+++ src/otx/algo/instance_segmentation/losses/__init__.py
@@ -5,8 +5,18 @@
 
 from .accuracy import accuracy
 from .dice_loss import DiceLoss
+from .hungarian_matcher import HungarianMatcher
+from .maskdino_criterion import SetCriterion
 from .roi_loss import ROICriterion
 from .rpn_loss import RPNCriterion
 from .rtmdet_inst_loss import RTMDetInstCriterion
 
-__all__ = ["accuracy", "DiceLoss", "RTMDetInstCriterion", "ROICriterion", "RPNCriterion"]
+__all__ = [
+    "accuracy",
+    "DiceLoss",
+    "RTMDetInstCriterion",
+    "ROICriterion",
+    "RPNCriterion",
+    "SetCriterion",
+    "HungarianMatcher",
+]